+++ blocks/form/block.ts
@@ -73,6 +73,11 @@ export type Field =
 
 export interface InputProps<T, F extends Field> {
   /**
+   * Whether or not the field should be disabled.
+   */
+  disabled?: boolean;
+
+  /**
    * A field error object.
    */
   error: string;
@@ -97,6 +102,10 @@ declare module '@appsemble/sdk' {
   interface Actions {
     onSubmitSuccess: {};
     onSubmit: {};
+  }
+
+  interface EventListeners {
+    data: {};
   }
 
   interface Parameters {