+++ blocks/form/src/components/StringInput/StringInput.tsx
@@ -11,6 +11,7 @@ type StringInputProps = InputProps<string, StringField>;
  * An input element for a text type schema.
  */
 export default function StringInput({
+  disabled,
   error,
   field: { format, label, maxLength, multiline, name, placeholder, readOnly, required },
   onInput,
@@ -18,6 +19,7 @@ export default function StringInput({
 }: StringInputProps): VNode {
   return (
     <Input
+      disabled={disabled}
       error={error && <FormattedMessage id="invalid" />}
       id={name}
       label={label || name}